backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.4
playwright==1.36.0
python-dotenv==1.0.0
sqlalchemy==2.0.19
PyYAML==6.0.1
pydantic==1.10.22
jinja2==3.1.2
cryptography==41.0.7 